add the controller and the service (#39)
* add the controller and the service

* Add `deleteCartById()` method in service and controller layers.

* solve merge issues

* add unit tests

* edit pom.xml

* fix linter issues

* fix tests

* remove checkstyle plugin

* fix tests

* use checkId method

* fix product tests

---------

+++ src/main/java/com/example/service/ProductService.java
@@ -1,6 +1,7 @@
 package com.example.service;
 
 import com.example.exception.InvalidActionException;
+import com.example.exception.ValidationException;
 import com.example.model.Product;
 import com.example.repository.CartRepository;
 import com.example.repository.OrderRepository;
@@ -73,12 +74,13 @@ public class ProductService extends MainService<Product, ProductRepository> {
 
     public void deleteProductById(final UUID productId) {
         checkId(productId);
+        if (productRepository.getProductById(productId) == null) {
+            throw new ValidationException("Product not found");
+        }
         if (orderRepository.isProductInOrder(productId)) {
             throw new InvalidActionException("Product is in active orders");
         }
 
-        // Delete the product from all carts that contain
-        // this product
         cartRepository.removeProductFromAllCarts(productId);
 
         productRepository.deleteProductById(productId);
